Imported Voicelines & Edited Menu
Voicelines for instructions on how to play, navigate, etc. and lines
like "Game Over" have been added. A few more will be added as needed as
beta finalizes. Some lines are extra in case and may or may not be used.

Tweaked main menu UI to be more in line of a radial-menu to improve
accessibility and to be morein the shape of a mobile screen. Added
colors and button highlights.