Added Voicelines to Arcade Mode and Minigame Buttons
Clicking Arcade Mode will play an intro description.
Hovering over each minigame button allows you to hear a description and
how to play the game.

Notes: Voicelines will be tweaked as Menu UI gets more figured out. (Do
we want to keep the "Begin Adventure" and "Load Game" buttons in for the
beta to demonstrate what we want to implement in the future, or should
we leave them out and just start on the "Arcade Mode" screen?)

Note2: Some discrepencies between the voice instructions and how menu
currently works because the desktop and mobile version will have to be
different. Will make voices match the beta/desktop version if that's
better.