Fixed High Score Screen + Text Updates
Was debugging it a while ago and forgot to change the code back; it
should update the scores and save them properly now

Increased/readjusted all text on the menu systems to be higher font and
not look fuzzy

Made score display say for example "Score: 5" instead of just "5", and
made larger.
Adjusted all score text size/format in all 3 games
Added "Lives: 3" text counter to survival game

Made Results text (incorrect/correct) on memory game bigger because it
was hard to read on mobile. Changed the wording to "Wrong direction!"
and "Correct!"

Made Memory game arrows a little bigger